+++ src/pages/LayoutBackofficePage.js
@@ -1,101 +1,49 @@
-import {Flex, Grid } from "@chakra-ui/react";
-import React from "react";
-import BackofficeCard from "../components/BackofficeCard";
-import novedadesIcon from "../assets/novedades.jpeg";
-import categoriasIcon from "../assets/categorias.png";
-import miembrosIcon from "../assets/miembos.png";
-import organizacionIcon from "../assets/organizacion.jpeg";
-import slidesIcon from "../assets/slides.png";
-import testimoniosIcon from "../assets/testimonios.png";
-import usuariosIcon from "../assets/usuarios.png";
-import actividadesIcon from "../assets/actividades.png";
+import React, {useState} from "react";
+import BackofficeUser from "../components/Backoffice/BackoficeUser";
+import BackofficeAdmin from "../components/Backoffice/BackofficeAdmin";
+
+
 
 
 
 const LayoutBackofficePage =() =>{
- const data = [
-        {
-          id: 1,
-          img: novedadesIcon,
-          alt: "novedades-icon",
-          title: "Novedades",
-          route:"/news"
-        },
-        {
-          id: 2,
-          img: actividadesIcon,
-          alt: "actividades-icon",
-          title: "Actividades",
-          route:"",
-        },
-        {
-          id: 3,
-          img: categoriasIcon,
-          alt: "categorias-icon",
-          title: "Categorias",
-          route:""
-        },
-        {
-          id: 4,
-          img: miembrosIcon,
-          alt: "miembros-icon",
-          title: "Miembros",
-          route:"",
-        },
-        {
-          id: 5,
-          img: organizacionIcon,
-          alt: "organizacion-icon",
-          title:"Organizacion",
-          route:"",
-        },
-        {
-          id: 6,
-          img: slidesIcon,
-          alt: "slides-icon",
-          title: "Slides",
-          route:"",
-        },
-        {
-          id: 7,
-          img: testimoniosIcon,
-          alt: "testimonios-icon",
-          title: "Testimonios",
-          route:"/testimonials",
-        },
-        {
-          id: 8,
-          img: usuariosIcon,
-          alt: "usuarioss-icon",
-          title: "Usuarios",
-          route:"",
-          
-        },
-    ];
 
+  const [status, setStatus] = useState(false)
 
- return(
-    <Flex
-    justifyContent={"center"}
-    >
-   <Grid
-   
-   gridTemplateColumns={["1fr 1fr","1fr 1fr","1fr 1fr 1fr 1fr"]}
-   gridGap={"10px"}
-   width={"80%"}
-   textAlign={"center"} //alinea los titulos
-   justifySelf={"center"}
-   >
-    {data.map((dato=>{
-        return(
-            <BackofficeCard key={dato.id} element={dato}/>
-        )
-    }))}
-   </Grid>
+  const authAdmin = ()=>{
+    var myHeaders = new Headers();
+    myHeaders.append("Authorization", `Bearer ${localStorage.getItem('token')}`);
+  
+   var requestOptions = {
+    method: 'GET',
+    headers: myHeaders,
+    redirect: 'follow'
+   };
+  
+   fetch(process.env.REACT_APP_SERVER_BASE_URL +"/auth/me", requestOptions)
+    .then(response => response.json())
+    .then(result => {
+      if(result.roleId ===1){
+        setStatus(true)
+      }
+    })
+    .catch(error => console.log('error', error));
+  }
+  
+  authAdmin()
 
-    </Flex>
+  
+  
+   if(status === false){
+    return(
+      <BackofficeUser></BackofficeUser>
 
- )
+    )
+   }else{
+    return(
+      <BackofficeAdmin></BackofficeAdmin>
+    )
+   }
 }
 
 export default LayoutBackofficePage;
